nextjs tanstack

+++ src/utils/helpers/fs/files.ts
@@ -1,7 +1,7 @@
 import { readFile, writeFile } from "fs/promises";
 import { printHelpers } from "../print-tools";
 import { existsSync } from "fs";
-
+import fsextra from "fs-extra";
 /**
  * Writes or overwrites a file with the given content.
  *
@@ -9,12 +9,25 @@ import { existsSync } from "fs";
  * @param {string} path - The path of the file to be written or overwritten.
  * @return {Promise<{success: boolean, operation: string}>} - A promise that resolves to an object indicating the success of the operation and the type of write operation performed.
  */
-export async function writeOrOverWriteFile(content: string, path: string) {
+export async function writeOrOverWriteFile(path: string, content: string) {
+  // printHelpers.info(path , content);
+  // printHelpers.info("content " + content);
   try {
-    const is_file_path = existsSync(content);
-    const write_content = is_file_path
-      ? await readFile(content, "utf-8")
-      : content;
+    const is_file_path = existsSync(path);
+    // printHelpers.info("writing file " + path);
+    if (!is_file_path) {
+      // printHelpers.warning("creating file " + path);
+      await fsextra.ensureFile(path);
+      await writeFile(path, content, { flag: "wx", encoding: "utf-8" }).catch((error: any) => {
+        printHelpers.error("error writing file " + error.message);
+        // throw error;
+      });
+      return {
+        success: true,
+        operation: "create file",
+      };
+    }
+    const write_content = await readFile(content, "utf-8");
     await writeFile(path, write_content);
     // printHelpers.success("file written successfully");
     return {
